requirements.txt
Flask==0.7.2
readability-lxml==0.2.3
lxml==2.3.2
chardet==1.0.1
